Update Search Component
I added a search message coming from the OMDb API
if the search term provided is too narrow for the api
to accept. I also tried to display the result of either
the message or the number of results being sent back,
but not both. There could be some errors popping up
but it works as is.